requirements.txt
<<<<<<< HEAD
setuptools
addict
aenum
aiohttp
anyio
appdirs
astunparse
blendmodes
clean-fid
easydev
extcolors
facexlib
filetype
future
gdown
gfpgan
GitPython
httpcore
=======
GitPython
Pillow
accelerate

basicsr
blendmodes
clean-fid
einops
fastapi>=0.90.1
gfpgan
gradio==3.41.2
>>>>>>> cf2772fab0af5573da775e7437e6acdca424f26e
inflection
jsonmerge
kornia
lark
<<<<<<< HEAD
lmdb
lpips
omegaconf
open-clip-torch
piexif
psutil
pyyaml
resize-right
rich
safetensors
scipy
tb_nightly
tensordict
toml
torchdiffeq
voluptuous
yapf
scikit-image
basicsr
fasteners
dctorch
pymatting
httpx==0.24.1
compel==2.0.2
torchsde==0.2.6
clip-interrogator==0.6.0
antlr4-python3-runtime==4.9.3
requests==2.31.0
tqdm==4.66.1
accelerate==0.25.0
opencv-contrib-python-headless==4.8.1.78
diffusers==0.25.0
einops==0.4.1
gradio==3.43.2
huggingface_hub==0.20.1
numexpr==2.8.4
numpy==1.24.4
numba==0.57.1
pandas==1.5.3
protobuf==3.20.3
pytorch_lightning==1.9.4
tokenizers==0.15.0
transformers==4.36.2
tomesd==0.1.3
urllib3==1.26.18
Pillow==10.1.0
timm==0.9.12
pydantic==1.10.13
typing-extensions==4.9.0
peft
=======
numpy
omegaconf
open-clip-torch

piexif
psutil
pytorch_lightning
realesrgan
requests
resize-right

safetensors
scikit-image>=0.19
timm
tomesd
torch
torchdiffeq
torchsde
transformers==4.30.2
>>>>>>> cf2772fab0af5573da775e7437e6acdca424f26e
